+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for Nautilus
 Fixed:
 ^^^^^^
 - Issues with liveview, cancelation and progress bar show hints
+- Uninitialized data for liveview when running without auto contrast/brightness
 
 
 0.1.21 (2023-04-24)